requirements.txt
fastapi
aiohttp
uvicorn
pydantic
requests
schedule
bs4
lxml
pytz
python-dotenv
httpx
redis
apscheduler
sqlalchemy
alembic